# backend/docs/swagger.yaml
definitions:
...
  api.Response:
    properties:
      data: {}
      error:
        type: string
      language:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  api.addScheduleRequest:
    properties:
      backup_type:
        type: string
      description:
        type: string
      id:
        type: string
      schedule:
        type: string
    required:
    - description
    - id
    - schedule
    type: object
...
  api.backupActivityItem:
    properties:
      action:
        description: backup, restore, cleanup
        type: string
      description:
        type: string
      duration:
        type: string
      filename:
        type: string
      size:
        type: integer
      status:
        description: success, failure, warning
        type: string
      timestamp:
        type: string
    type: object
  api.backupConfigSummary:
    properties:
      compress_backups:
        type: boolean
      encrypt_backups:
        type: boolean
      max_backup_count:
        type: integer
      retention_days:
        type: integer
      storage_type:
        type: string
      validate_backups:
        type: boolean
    type: object
  api.backupHealthStatus:
    properties:
      issues:
        items:
          type: string
        type: array
      last_check:
        type: string
      overall:
        description: healthy, warning, critical
        type: string
    type: object
  api.backupHistoryItem:
    properties:
      description:
        type: string
      duration:
        type: string
      error:
        type: string
      filename:
        type: string
      operation:
        type: string
      size:
        type: integer
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  api.backupHistoryResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/api.backupHistoryItem'
        type: array
      total:
        type: integer
    type: object
...
  api.backupStatusResponse:
    properties:
      auto_backup:
        type: boolean
      configuration:
        $ref: '#/definitions/api.backupConfigSummary'
      enabled:
        type: boolean
      health:
        $ref: '#/definitions/api.backupHealthStatus'
      last_backup:
        type: string
      next_backup:
        type: string
      oldest_backup:
        type: string
      recent_activity:
        items:
          $ref: '#/definitions/api.backupActivityItem'
        type: array
      total_backups:
        type: integer
      total_size:
        type: integer
    type: object
  api.backupValidationResponse:
    properties:
      checksum_match:
        type: boolean
      file_exists:
        type: boolean
      filename:
        type: string
      issues:
        items:
          type: string
        type: array
      last_modified:
        type: string
      readable_file:
        type: boolean
      size:
        type: integer
      valid:
        type: boolean
      valid_structure:
        type: boolean
      warnings:
        items:
          type: string
        type: array
    type: object
...
  api.cleanupRequest:
    properties:
      max_age:
        description: Duration string like "720h" for 30 days
        type: string
    type: object
...
  api.enhancedBackupRequest:
    properties:
      description:
        type: string
      type:
        description: manual, automatic, migration, etc.
        type: string
    required:
    - description
    - type
    type: object
  api.enhancedBackupResponse:
    properties:
      checksum:
        type: string
      compressed:
        type: boolean
      download_url:
        type: string
      duration:
        type: string
      encrypted:
        type: boolean
      filename:
        type: string
      size:
        type: integer
      success:
        type: boolean
      validated:
        type: boolean
      warnings:
        items:
          type: string
        type: array
    type: object
  api.enhancedRestoreRequest:
    properties:
      filename:
        type: string
    required:
    - filename
    type: object
  api.enhancedRestoreResponse:
    properties:
      duration:
        type: string
      records_count:
        type: integer
      success:
        type: boolean
      tables_count:
        type: integer
      warnings:
        items:
          type: string
        type: array
    type: object
...
  api.scheduleInfo:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      last_run:
        type: string
      next_run:
        type: string
      schedule:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/backups/{filename}/validate:
    post:
      consumes:
      - application/json
      description: Validates the integrity and structure of a backup file
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Backup validation completed
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.backupValidationResponse'
              type: object
        "400":
          description: Invalid filename
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Backup file not found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Validate backup file
      tags:
      - admin
  /api/v1/admin/backups/cleanup:
    post:
      consumes:
      - application/json
      description: Manually trigger cleanup of old backup files
      parameters:
      - description: Cleanup parameters
        in: body
        name: cleanup
        schema:
          $ref: '#/definitions/api.cleanupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup completed successfully
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to cleanup backups
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Manual cleanup of old backups
      tags:
      - admin
...
  /api/v1/admin/backups/enhanced:
    post:
      consumes:
      - application/json
      description: Creates a backup with advanced features like compression, encryption,
        and validation
      parameters:
      - description: Enhanced backup details
        in: body
        name: backup
        required: true
        schema:
          $ref: '#/definitions/api.enhancedBackupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Backup created successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.enhancedBackupResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to create backup
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create enhanced database backup
      tags:
      - admin
  /api/v1/admin/backups/enhanced/restore:
    post:
      consumes:
      - application/json
      description: Restores database with advanced validation and safety measures
      parameters:
      - description: Enhanced restore details
        in: body
        name: restore
        required: true
        schema:
          $ref: '#/definitions/api.enhancedRestoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Database restored successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.enhancedRestoreResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Backup file not found
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to restore database
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Restore database from enhanced backup
      tags:
      - admin
  /api/v1/admin/backups/history:
    get:
      description: Get backup history and statistics
      parameters:
      - default: 50
        description: Limit number of results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Backup history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.backupHistoryResponse'
              type: object
        "500":
          description: Failed to get backup history
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get backup history
      tags:
      - admin
...
  /api/v1/admin/backups/schedules:
    get:
      description: Get all active backup schedules
      produces:
      - application/json
      responses:
        "200":
          description: Backup schedules retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api.scheduleInfo'
                  type: array
              type: object
        "500":
          description: Failed to get schedules
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get backup schedules
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Add a new backup schedule
      parameters:
      - description: Schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/api.addScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule added successfully
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to add schedule
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Add backup schedule
      tags:
      - admin
  /api/v1/admin/backups/schedules/{id}:
    delete:
      description: Remove a backup schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule removed successfully
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Failed to remove schedule
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove backup schedule
      tags:
      - admin
  /api/v1/admin/backups/status:
    get:
      consumes:
      - application/json
      description: Returns backup system status, recent activity, and health metrics
      produces:
      - application/json
      responses:
        "200":
          description: Backup status retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.backupStatusResponse'
              type: object
        "500":
          description: Failed to retrieve backup status
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get backup status and health
      tags:
      - admin
...
swagger: "2.0"